Added in YAML config, implemented CLI for paths
Most parameters now controlled by an external config file. Still some
parameters to switch over to config but this will require a more
extensive rework of the `train_model.py` script.

The training script now takes command line arguments for paths to data
files, config files, and requires a directory to store results. This
makes the code more generalisable and usable on other machines (with the
view to running on HPC in the future).

Further, there is a key success in this commit - the implementation of
cyclic learning rate with the Adam optimiser led to some improved
results. This is to be explored further.