Plots verfied and corrected
An issue was present wherein the entire datasets where being plotted
rather than the validation or test portion. This was due to the way the
dataset was being accessed; the subset object after `random_split` has
access to the parent dataset and thus this was being plotted. This has
been corrected, and now the validation plots have the correct number of
data points, as do the eval plots.

Resolves: #43